# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction Service API",
        "description": "API de gestion des transactions",
        "contact": {
            "email": "n.taffot@elyft.tech"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8003",
            "description": "Transaction Service Server"
        }
    ],
    "paths": {
...
        "/api/transactions/operations": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Liste des transactions",
                "description": "Récupère la liste paginée des transactions avec filtres",
                "operationId": "c1be3ced5f0c40d6604343c109bd57f8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "FAILED",
                                "CANCELLED",
                                "EXPIRED",
                                "SUCCESSFULL"
                            ]
                        }
                    },
                    {
                        "name": "transaction_type",
                        "in": "query",
                        "description": "Filtrer par type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "deposit",
                                "withdrawal"
                            ]
                        }
                    },
                    {
                        "name": "customer_phone",
                        "in": "query",
                        "description": "Filtrer par téléphone client",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Date de début (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Date de fin (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Créer une nouvelle transaction",
                "description": "Crée une nouvelle transaction de dépôt ou retrait",
                "operationId": "55a58cbc098f861891b8e6343b78fdc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "entreprise_id",
                                    "wallet_id",
                                    "operator_id",
                                    "transaction_type",
                                    "amount",
                                    "currency_code",
                                    "customer_phone"
                                ],
                                "properties": {
                                    "entreprise_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "wallet_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "operator_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "transaction_type": {
                                        "type": "string",
                                        "enum": [
                                            "deposit",
                                            "withdrawal"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 1
                                    },
                                    "currency_code": {
                                        "type": "string",
                                        "maxLength": 3,
                                        "minLength": 3,
                                        "example": "XAF"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "237600000000"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "api_key_used": {
                                        "type": "string"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/transactions/operations/status/{status}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Transactions par statut",
                "description": "Récupère les transactions d'un statut donné",
                "operationId": "22b8f6e0792f6f0bb64bb94a5b01a0db",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "FAILED",
                                "CANCELLED",
                                "EXPIRED",
                                "SUCCESSFULL"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/operations/{id}/retry": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Relancer une transaction",
                "description": "Relance une transaction échouée",
                "operationId": "470803d8598623cea52e25658a7e8305",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction relancée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Transaction": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "entreprise_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "wallet_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "operator_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "transaction_type": {
                        "type": "string",
                        "enum": [
                            "deposit",
                            "withdrawal"
                        ]
                    },
                    "amount": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "currency_code": {
                        "type": "string"
                    },
                    "operator_commission": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "internal_commission": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "net_amount": {
                        "type": "number",
                        "format": "decimal"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "FAILED",
                            "CANCELLED",
                            "EXPIRED",
                            "SUCCESSFULL"
                        ]
                    },
                    "operator_status": {
                        "type": "string",
                        "nullable": true
                    },
                    "operator_transaction_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "customer_phone": {
                        "type": "string"
                    },
                    "customer_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "initiated_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "datetime",
                        "nullable": true
                    },
                    "api_key_used": {
                        "type": "string",
                        "nullable": true
                    },
                    "ip_address": {
                        "type": "string",
                        "nullable": true
                    },
                    "user_agent": {
                        "type": "string",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "failure_reason": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}